+++ ReleaseNotes.md
@@ -26,3 +26,4 @@
 - Add MileLoadLibraryFromSystem32 function.
 - Add Mile::WinRT::EmptyValue template function.
 - Add Mile::WinRT::Property template class.
+- Add Mile::WinRT::Event template class.